+++ .travis.yml
@@ -1,8 +1,3 @@
 language: rust
 rust:
-  - stable
-  - beta
   - nightly
-matrix:
-  allow_failures:
-    - rust: nightly
