+++ Cargo.toml
@@ -5,6 +5,7 @@ authors = ["Nathan Hoad <nathan@getoffmalawn.com>"]
 
 [dependencies]
 httparse = "1.1.2"
+mioco = "0.5.0"
 
 [dependencies.url]
 git = "https://github.com/servo/rust-url"
