+++ src/lib.rs
@@ -3,3 +3,4 @@ pub mod macros;
 
 pub mod headers;
 pub mod request;
+pub mod server;
